+++ ouzel/core/Window.cpp
@@ -28,7 +28,7 @@
 namespace ouzel
 {
     Window::Window(Engine& initEngine,
-                   const Size2<float>& newSize,
+                   const Size2<uint32_t>& newSize,
                    bool newResizable,
                    bool newFullscreen,
                    bool newExclusiveFullscreen,
@@ -206,7 +206,7 @@ namespace ouzel
         nativeWindow->addCommand(command);
     }
 
-    void Window::setSize(const Size2<float>& newSize)
+    void Window::setSize(const Size2<uint32_t>& newSize)
     {
         if (size != newSize)
         {
