+++ ouzel/graphics/metal/MetalRenderDevice.mm
@@ -132,7 +132,7 @@ namespace ouzel
         }
 
         void MetalRenderDevice::init(Window* newWindow,
-                                     const Size2<float>& newSize,
+                                     const Size2<uint32_t>& newSize,
                                      uint32_t newSampleCount,
                                      Texture::Filter newTextureFilter,
                                      uint32_t newMaxAnisotropy,
@@ -220,7 +220,7 @@ namespace ouzel
             renderPassDescriptor.depthAttachment.clearDepth = clearDepth;
         }
 
-        void MetalRenderDevice::setSize(const Size2<float>& newSize)
+        void MetalRenderDevice::setSize(const Size2<uint32_t>& newSize)
         {
             RenderDevice::setSize(newSize);
