+++ ouzel/graphics/opengl/linux/OGLRenderDeviceLinux.hpp
@@ -29,11 +29,11 @@ namespace ouzel
             OGLRenderDeviceLinux(const std::function<void(const Event&)>& initCallback);
             ~OGLRenderDeviceLinux();
 
-            std::vector<Size2<float>> getSupportedResolutions() const override;
+            std::vector<Size2<uint32_t>> getSupportedResolutions() const override;
 
         private:
             void init(Window* newWindow,
-                      const Size2<float>& newSize,
+                      const Size2<uint32_t>& newSize,
                       uint32_t newSampleCount,
                       Texture::Filter newTextureFilter,
                       uint32_t newMaxAnisotropy,
